% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{apply_filters}
\alias{apply_filters}
\title{Filter AOIs based on thresholds set by set_thresholds function}
\usage{
apply_filters(ds)
}
\arguments{
\item{ds}{list dataset produced by set_thresholds}
}
\value{
list dataset
}
\description{
Filter AOIs based on thresholds set by set_thresholds function
}
\examples{
data(example_ds, package= "stgeomx")
example_ds <- preprocess(example_ds)
example_ds <- set_thresholds(example_ds)
apply_filters(example_ds)

}
