% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{background_subtract}
\alias{background_subtract}
\title{subtract background noise from each sample}
\usage{
background_subtract(x, offset = 0, min.count = 1)
}
\arguments{
\item{x}{count matrix}

\item{offset}{vector of bg noise levels per sample}

\item{min.count}{minimum count threshold}
}
\value{
matrix of background-subtracted counts
}
\description{
subtract background noise from each sample
}
