% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgcorrect.R
\name{bgcorrect}
\alias{bgcorrect}
\title{Background correction}
\usage{
bgcorrect(ds, method = c("qq", "bgsub", "none"), bw.adjust = 1, bg.quant = 0.5)
}
\arguments{
\item{ds}{stgeomx dataset}

\item{method}{correction method to choose from}

\item{bw.adjust}{bandwidth adjustment factor}

\item{bg.quant}{background subtraction quantile}
}
\value{
stgeomx dataset
}
\description{
Background correction
}
\examples{
data(example_ds, package= "stgeomx")
bgcorrect(example_ds, "qq")
bgcorrect(example_ds, "bgsub")
bgcorrect(example_ds, "none")

}
