% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgcorrect.R
\name{bgcorrect_qq}
\alias{bgcorrect_qq}
\title{background correction using quantile-quantile approach}
\usage{
bgcorrect_qq(x, bg, bw.adjust = 1, bg.quant = 0.5)
}
\arguments{
\item{x}{matrix of counts}

\item{bg}{vector of TRUE/FALSE corresponding to rows of x}

\item{bw.adjust}{bandwidth adjustment factor for density function}

\item{bg.quant}{subtract background quantile from all genes}
}
\value{
list where element 'y' contains corrected counts
}
\description{
background correction using quantile-quantile approach
}
