% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{calc_bg_quantiles}
\alias{calc_bg_quantiles}
\title{Compute probe quantiles relative to background probe ecdf}
\usage{
calc_bg_quantiles(ds)
}
\arguments{
\item{ds}{dataset}
}
\value{
matrix of quantiles corresponding to genes (rows) and aois (cols)
}
\description{
Compute probe quantiles relative to background probe ecdf
}
