% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{calc_bg_stats}
\alias{calc_bg_stats}
\title{Compute qc metrics}
\usage{
calc_bg_stats(meta, counts, bg_lod_quantile)
}
\arguments{
\item{meta}{tibble of count metadata}

\item{counts}{tibble of counts}

\item{bg_lod_quantile}{number range (0.0-1.0)}
}
\description{
Compute qc metrics
}
