% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{calc_norm_factors_rle}
\alias{calc_norm_factors_rle}
\title{edgeR package
https://rdrr.io/bioc/edgeR/src/R/calcNormFactors.R
Cite Mark Robinson
Scale factors as in Anders et al (2010)
Mark Robinson
Created 16 Aug 2010}
\usage{
calc_norm_factors_rle(data)
}
\arguments{
\item{data}{matrix of counts}
}
\description{
edgeR package
https://rdrr.io/bioc/edgeR/src/R/calcNormFactors.R
Cite Mark Robinson
Scale factors as in Anders et al (2010)
Mark Robinson
Created 16 Aug 2010
}
