% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{calc_norm_factors_tmm}
\alias{calc_norm_factors_tmm}
\title{edgeR package
TMM between two libraries
Cite Mark Robinson
Copied from https://rdrr.io/bioc/edgeR/src/R/calcNormFactors.R}
\usage{
calc_norm_factors_tmm(
  x,
  logratioTrim = 0.3,
  sumTrim = 0.05,
  doWeighting = TRUE,
  Acutoff = -1e+10
)
}
\arguments{
\item{x}{count matrix}

\item{logratioTrim}{see edgeR documentation}

\item{sumTrim}{see edgeR documentation}

\item{doWeighting}{see edgeR documentation}

\item{Acutoff}{see edgeR documentation}
}
\description{
edgeR package
TMM between two libraries
Cite Mark Robinson
Copied from https://rdrr.io/bioc/edgeR/src/R/calcNormFactors.R
}
