% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{calc_qc_metrics}
\alias{calc_qc_metrics}
\title{Calculate QC metrics
adds qc columns to the samples tibble}
\usage{
calc_qc_metrics(samples, meta, counts, bg_lod_quantile)
}
\arguments{
\item{samples}{tibble of sample information}

\item{meta}{tibble of gene metadata}

\item{counts}{tibble of count data}

\item{bg_lod_quantile}{number (0.0-1.0)}
}
\value{
samples tibble with added qc columns
}
\description{
Calculate QC metrics
adds qc columns to the samples tibble
}
