% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{calc_snauc}
\alias{calc_snauc}
\title{Calculate gene vs background AUC (snAUC) for each AOI}
\usage{
calc_snauc(meta, counts)
}
\arguments{
\item{meta}{tibble count metadata}

\item{counts}{tibble counts}
}
\value{
vector of AUC values for each AOI
}
\description{
Calculate gene vs background AUC (snAUC) for each AOI
}
