% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{compute_quantiles_kde}
\alias{compute_quantiles_kde}
\title{Compute quantiles of values in vector 'x' along
compute ecdf of background distribution and use it to
interpolate quantiles of genes in vector}
\usage{
compute_quantiles_kde(x, bg, bw.adjust = 2)
}
\arguments{
\item{x}{vector of numeric values}

\item{bg}{vector of TRUE/FALSE corresponding to x}

\item{bw.adjust}{bandwidth adjustment multiplier for 'density' function}
}
\value{
vector of quantiles corresponding to array 'x'
}
\description{
Compute quantiles of values in vector 'x' along
compute ecdf of background distribution and use it to
interpolate quantiles of genes in vector
}
