% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{copy}
\alias{copy}
\title{copy dataset}
\usage{
copy(ds)
}
\arguments{
\item{ds}{stgeomx dataset}
}
\value{
dataset
}
\description{
copy dataset
}
\examples{
data(example_ds, package = "stgeomx")
duplicate <- copy(example_ds)

}
