% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{cpm}
\alias{cpm}
\title{convert to counts per million}
\usage{
cpm(x)
}
\arguments{
\item{x}{tibble/dataframe/matrix of counts}
}
\value{
matrix counts-per-million
}
\description{
convert to counts per million
}
