% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{cpm2counts}
\alias{cpm2counts}
\title{convert from cpm to counts}
\usage{
cpm2counts(x, tot_counts)
}
\arguments{
\item{x}{tibble/dataframe/matrix of cpm}

\item{tot_counts}{vector of total counts per sample}
}
\value{
matrix counts
}
\description{
convert from cpm to counts
}
