% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{dotplot_aoi_bg_vs_signal}
\alias{dotplot_aoi_bg_vs_signal}
\title{compare median gene to bg cpm}
\usage{
dotplot_aoi_bg_vs_signal(x)
}
\arguments{
\item{x}{list dataset}
}
\value{
ggplot object
}
\description{
compare median gene to bg cpm
}
