% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg.R
\name{get_hvg_tbl}
\alias{get_hvg_tbl}
\title{Model and rank highly variable genes in dataset}
\usage{
get_hvg_tbl(x, loess.span = 0.25, min_var_quantile = 0.01)
}
\arguments{
\item{x}{matrix/data frame of normalized log2 transformed gene expression,
where gene identifiers are present in the row names of the matrix}

\item{loess.span}{span parameter for loess model}

\item{min_var_quantile}{[0.0-1.0] used as lower bound for predicted
variance by loess model (in rare cases the predicted variance by loess
can be less than zero or very close to zero, resulting in high variance
ratios)}
}
\value{
tibble object with per-gene parameters
}
\description{
Models mean-variance relationship using loess. Ratio of actual variance
versus modeled variance is used to rank genes
}
