% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{init}
\alias{init}
\title{initialize dataset}
\usage{
init(samples, meta, counts, x)
}
\arguments{
\item{samples}{tibble with sample information}

\item{meta}{tibble with gene metadata}

\item{counts}{tibble with gene count data}

\item{x}{tibble with gene count data}
}
\value{
dataset
}
\description{
initialize dataset
}
\examples{
samples <- data.frame(
  SlideName = c("brain_001", "brain_002"),
  ScanLabel = c("Brain 1", "Brain 2"),
  aoi = c("s01_r001_fullroi", "s02_r001_fullroi")
)
meta <- data.frame(
  probe = c("example01", "NegProbe-WTX01"),
  gene = c("example", "NegProbe-WTX"),
  bg = c(FALSE, TRUE)
)
counts <- data.frame(
  s01_r001_fullroi <- c(100, 32),
  s02_r001_fullroi <- c(95, 24)
)
x <- data.frame(
  s01_r001_fullroi <- c(100, 32),
  s02_r001_fullroi <- c(95, 24)
)
new_ds <- init(samples, meta, counts, x)

}
