% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{merge}
\alias{merge}
\title{Merge two datasets}
\usage{
merge(a, b)
}
\arguments{
\item{a}{dataset produced by st_geomx_read}

\item{b}{dataset produced by st_geomx_read}
}
\value{
dataset
}
\description{
This function should only be used to merge datasets that are generated
from the same probe kit.
}
\details{
This function can join datasets with two options: union and intersect.
Intersect will keep columns that are common to both datasets,
whereas union will keep all columns.
}
\examples{
# example code
data(example_ds, package= "stgeomx")
merge_ds(example_ds, example_ds)

}
