% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{merge_probes}
\alias{merge_probes}
\title{Merge multiple probes per gene}
\usage{
merge_probes(ds)
}
\arguments{
\item{ds}{list produced by st_geomx_process}
}
\value{
list
}
\description{
Merge multiple probes per gene
}
\examples{
data(example_ds, package= "stgeomx")
example_ds <- preprocess(example_ds)
merge_probes(example_ds)

}
