% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{normalize_quantile}
\alias{normalize_quantile}
\title{quantile normalization}
\usage{
normalize_quantile(x)
}
\arguments{
\item{x}{matrix of numeric values}
}
\value{
matrix of quantile normalized counts
}
\description{
quantile normalization
}
