% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{plot_aoi_filter}
\alias{plot_aoi_filter}
\title{Set of plots to guide aoi filtering}
\usage{
plot_aoi_filter(ds, min_counts = 0, min_auc = 0.5, min_frac_expr = 0)
}
\arguments{
\item{ds}{dataset}

\item{min_counts}{number counts}

\item{min_auc}{number minimum area under the signal-noise ROC curve (snauc)}

\item{min_frac_expr}{number threshold range (0-1.0)}
}
\value{
ggplot object
}
\description{
Set of plots to guide aoi filtering
}
\examples{
data(example_ds, package = "stgeomx")
example_ds <- preprocess(example_ds)
plot_aoi_filter(example_ds)

}
