% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{plot_aoi_qc}
\alias{plot_aoi_qc}
\title{QC plots for background correction}
\usage{
plot_aoi_qc(ds, aoi, bw.adjust = 2, bg.quant = 0.5, bg.lod.quant = 0.9)
}
\arguments{
\item{ds}{dataset}

\item{aoi}{name of aoi}

\item{bw.adjust}{integer adjustment factor for bandwidth function}

\item{bg.quant}{background quantile to subtract to the counts}

\item{bg.lod.quant}{limit of detection quantile}
}
\value{
ggplot object
}
\description{
QC plots for background correction
}
\examples{
data(example_ds, package = "stgeomx")
example_ds <- preprocess(example_ds)
plot_aoi_qc(example_ds, "s01_r001_fullroi")

}
