% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{plot_bgcorrect}
\alias{plot_bgcorrect}
\title{qc plot of background noise (x axis) versus gene quantiles (y axis)
uses sample metadata column 'keep' to filter aois before plotting
uses gene metadata column 'keep' to filter genes before plotting}
\usage{
plot_bgcorrect(
  ds,
  logscale = TRUE,
  quantiles = c(0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99)
)
}
\arguments{
\item{ds}{list dataset}

\item{logscale}{TRUE for log scale, FALSE for linear scale}

\item{quantiles}{vector of quantiles [0.0-1.0]}
}
\value{
ggplot object
}
\description{
qc plot of background noise (x axis) versus gene quantiles (y axis)
uses sample metadata column 'keep' to filter aois before plotting
uses gene metadata column 'keep' to filter genes before plotting
}
\examples{
data(example_ds, package = "stgeomx")
example_ds <- preprocess(example_ds)
plot_bgcorrect(example_ds)

}
