% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{plot_bgcorrect1}
\alias{plot_bgcorrect1}
\title{qc plot to assess relationship of bg versus gene}
\usage{
plot_bgcorrect1(ds, quantiles = c(0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99))
}
\arguments{
\item{ds}{list dataset}

\item{quantiles}{vector of quantiles [0.0-1.0]}
}
\value{
ggplot object
}
\description{
qc plot to assess relationship of bg versus gene
}
\examples{
data(example_ds, package = "stgeomx")
example_ds <- preprocess(example_ds)
plot_bgcorrect1(example_ds)

}
