% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{plot_bgcorrect3}
\alias{plot_bgcorrect3}
\title{plot total background noise (x axis) versus gene quantiles (y axis)}
\usage{
plot_bgcorrect3(ds, quantiles = c(0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99))
}
\arguments{
\item{ds}{list dataset}

\item{quantiles}{vector of quantiles [0.0-1.0]}
}
\value{
ggplot object
}
\description{
plot total background noise (x axis) versus gene quantiles (y axis)
}
\examples{
data(example_ds, package = "stgeomx")
example_ds <- preprocess(example_ds)
plot_bgcorrect3(example_ds)

}
