% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{plot_expr_dist}
\alias{plot_expr_dist}
\title{boxplot showing distribution of count values}
\usage{
plot_expr_dist(ds, bg.lod.quantile = 0.9)
}
\arguments{
\item{ds}{list dataset}

\item{bg.lod.quantile}{number between 0 and 1}
}
\value{
ggplot object
}
\description{
boxplot showing distribution of count values
}
\examples{
data(example_ds, package = "stgeomx")
example_ds <- preprocess(example_ds)
plot_expr_dist(example_ds)

}
