% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{plot_gene_filter}
\alias{plot_gene_filter}
\title{Set of plots to guide gene filtering}
\usage{
plot_gene_filter(ds, min_frac_expr = 0)
}
\arguments{
\item{ds}{list dataset}

\item{min_frac_expr}{number threshold range (0-1.0)}
}
\value{
ggplot object
}
\description{
Set of plots to guide gene filtering
}
\examples{
data(example_ds, package = "stgeomx")
example_ds <- preprocess(example_ds)
plot_gene_filter(example_ds)

}
