% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg.R
\name{plot_hvg_model}
\alias{plot_hvg_model}
\title{plot of gene mean versus variance with line showing predicted (expected)
variance by loess model. dashed line shows lower bound to predicted}
\usage{
plot_hvg_model(
  x,
  nhvg = 500,
  nlabel = 100,
  loess.span = 0.25,
  min_var_quantile = 0.01
)
}
\arguments{
\item{x}{matrix/data frame (normalized log2 gene expression)}

\item{nhvg}{number of top highly variable genes to color differently}

\item{nlabel}{number of top highly variable genes to label}

\item{loess.span}{span parameter for loess model}

\item{min_var_quantile}{[0.0-1.0] variance quantile to use as lower bound}
}
\value{
ggplot object
}
\description{
plot of gene mean versus variance with line showing predicted (expected)
variance by loess model. dashed line shows lower bound to predicted
}
