% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{prepare_input}
\alias{prepare_input}
\title{Standardize names using Slide, ROI, Segment, and AOI columns of sample
annotation table. Partition count table into metadata and counts.}
\usage{
prepare_input(
  samples,
  counts,
  slide_key_col,
  roi_key_col,
  segment_key_col,
  aoi_key_col,
  probe_key_col,
  gene_key_col,
  negprobe_regex
)
}
\arguments{
\item{samples}{tibble of sample information}

\item{counts}{tibble of count data}

\item{slide_key_col}{sample column name containing scan}

\item{roi_key_col}{sample column name containing roi}

\item{segment_key_col}{sample column name containing segment}

\item{aoi_key_col}{sample column name corresponding to columns of count sheet}

\item{probe_key_col}{count sheet column name unique probe identifier}

\item{gene_key_col}{count sheet column name unique gene identifier}

\item{negprobe_regex}{regular expression corresponding to negative probes}
}
\value{
ds
}
\description{
Standardize names using Slide, ROI, Segment, and AOI columns of sample
annotation table. Partition count table into metadata and counts.
}
