% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess input data}
\usage{
preprocess(ds, bg_lod_quantile = 0.9)
}
\arguments{
\item{ds}{dataset produced by st_geomx_read}

\item{bg_lod_quantile}{number (0.0-1.0)}
}
\value{
dataset
}
\description{
Preprocess input data
}
\examples{
data(example_ds, package= "stgeomx")
preprocess(example_ds)

data(example_ds, package= "stgeomx")
preprocess(example_ds, 0.8)

}
