% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{qnorm}
\alias{qnorm}
\title{quantile normalization}
\usage{
qnorm(x)
}
\arguments{
\item{x}{matrix of numeric values}
}
\value{
matrix of quantile normalized relative counts
}
\description{
ties are handled using the "max" method (highest rank)
values are shifted to have a minimum of zero and renormalized such that
columns sum to 1 million
}
