% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_tsv}
\alias{read_tsv}
\title{Read Nanostring GeoMx DSP data from tab-delimited text files}
\usage{
read_tsv(
  sample_tsv_file,
  count_tsv_file,
  slide_key_col = "ScanLabel",
  roi_key_col = "ROILabel",
  segment_key_col = "SegmentLabel",
  aoi_key_col = "SegmentDisplayName",
  probe_key_col = "ProbeDisplayName",
  gene_key_col = "TargetName",
  negprobe_regex = "^NegProbe"
)
}
\arguments{
\item{sample_tsv_file}{tab-delimited text file with sample annotations}

\item{count_tsv_file}{tab-delimited text file with count matrix/metadata}

\item{slide_key_col}{sample column name containing scan}

\item{roi_key_col}{sample column name containing roi}

\item{segment_key_col}{sample column name containing segment}

\item{aoi_key_col}{sample column name corresponding to columns of count sheet}

\item{probe_key_col}{count sheet column name unique probe identifier}

\item{gene_key_col}{count sheet column name unique gene identifier}

\item{negprobe_regex}{regular expression corresponding to negative probes}
}
\value{
dataset
}
\description{
Read Nanostring GeoMx DSP data from tab-delimited text files
}
\examples{
sample_tsv <- system.file("extdata", "simple_geomx_sample.tsv", package = "stgeomx")
count_tsv <- system.file("extdata", "simple_geomx_count.tsv", package = "stgeomx")
read_tsv(sample_tsv, count_tsv)

}
