% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_xlsx}
\alias{read_xlsx}
\title{Read Nanostring GeoMx DSP data from an Excel (XLSX) file}
\usage{
read_xlsx(
  xlsx_file,
  sample_sheet = "SegmentProperties",
  count_sheet = "BioProbeCountMatrix",
  slide_key_col = "ScanLabel",
  roi_key_col = "ROILabel",
  segment_key_col = "SegmentLabel",
  aoi_key_col = "SegmentDisplayName",
  probe_key_col = "ProbeDisplayName",
  gene_key_col = "TargetName",
  negprobe_regex = "^NegProbe"
)
}
\arguments{
\item{xlsx_file}{Excel file containing input data}

\item{sample_sheet}{Excel sheet name containing sample information}

\item{count_sheet}{Excel sheet name containing counts}

\item{slide_key_col}{sample sheet column name containing scan}

\item{roi_key_col}{sample sheetcolumn name containing roi}

\item{segment_key_col}{sample sheet column name containing segment}

\item{aoi_key_col}{sample sheet column name corresponding to columns of count sheet}

\item{probe_key_col}{count sheet column name unique probe identifier}

\item{gene_key_col}{count sheet column name unique gene identifier}

\item{negprobe_regex}{regex corresponding to negative probes}
}
\value{
list containing samples, metadata, and counts
}
\description{
Read Nanostring GeoMx DSP data from an Excel (XLSX) file
}
\examples{
xlsx_file <- system.file("extdata", "simple_geomx.xlsx", package = "stgeomx")
read_xlsx(xlsx_file)

}
