% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{remove_empty_aois}
\alias{remove_empty_aois}
\title{Remove AOIs with no valid counts}
\usage{
remove_empty_aois(ds, min_count = 1)
}
\arguments{
\item{ds}{list produced by st_geomx_merge_probes}

\item{min_count}{integer count threshold for removing aoi}
}
\value{
dataset
}
\description{
Remove AOIs with no valid counts
}
