% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{scatter_aoi_bg_vs_signal}
\alias{scatter_aoi_bg_vs_signal}
\title{scatter plot signal vs bg}
\usage{
scatter_aoi_bg_vs_signal(x)
}
\arguments{
\item{x}{list dataset}
}
\value{
ggplot object
}
\description{
scatter plot signal vs bg
}
