% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{set_thresholds}
\alias{set_thresholds}
\title{Apply user-defined thresholds to AOIs in preparation of filtering}
\usage{
set_thresholds(
  ds,
  min_counts = 0,
  min_auc = 0.5,
  aoi_min_frac_expr = 0,
  gene_min_frac_expr = 0
)
}
\arguments{
\item{ds}{list produced by st_geomx_merge_probes}

\item{min_counts}{integer minimum counts per aoi}

\item{min_auc}{number minimum snAUC per aoi}

\item{aoi_min_frac_expr}{number minimum frac of genes expressed per aoi}

\item{gene_min_frac_expr}{number minimum frac of samples expressing gene}
}
\value{
list dataset
}
\description{
This function does not perform filtering. It sets the 'keep' column in the
sample and gene metadata allowing users to generate plots of different
thresholds.
}
\details{
Sequencing depth and signal-to-noise ratio can vary dramatically across
experiments. The default parameter values will not perform any filtering.
We recommend using the plotting functions to aid in choosing
appropriate parameter values.
}
\examples{
data(example_ds, package= "stgeomx")
example_ds <- preprocess(example_ds)
set_thresholds(example_ds, 50000, 0.6, 0.2, 0.15)

}
