% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{st_geomx_filter}
\alias{st_geomx_filter}
\title{Filter AOIs based on thresholds set by st_geomx_filter_thresholds}
\usage{
st_geomx_filter(ds)
}
\arguments{
\item{ds}{list dataset produced by st_geomx_filter_thresholds}
}
\value{
list dataset
}
\description{
Filter AOIs based on thresholds set by st_geomx_filter_thresholds
}
