% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{st_geomx_merge_probes}
\alias{st_geomx_merge_probes}
\title{Merge multiple probes per gene}
\usage{
st_geomx_merge_probes(ds)
}
\arguments{
\item{ds}{list produced by st_geomx_process}
}
\value{
list
}
\description{
Merge multiple probes per gene
}
