% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{st_geomx_plot_aoi_filter}
\alias{st_geomx_plot_aoi_filter}
\title{Set of plots to guide aoi filtering}
\usage{
st_geomx_plot_aoi_filter(ds, min_counts = 0, min_auc = 0.5, min_frac_expr = 0)
}
\arguments{
\item{min_counts}{number counts}

\item{min_auc}{number minimum area under the signal-noise ROC curve (snauc)}

\item{min_frac_expr}{number threshold range (0-1.0)}

\item{list}{dataset}
}
\value{
ggplot object
}
\description{
Set of plots to guide aoi filtering
}
