% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{st_geomx_plot_bgsub}
\alias{st_geomx_plot_bgsub}
\title{qc plots to assess the effect of background subtraction}
\usage{
st_geomx_plot_bgsub(ds)
}
\arguments{
\item{list}{dataset}
}
\value{
ggplot object
}
\description{
qc plots to assess the effect of background subtraction
}
