% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{st_geomx_plot_count_distribution}
\alias{st_geomx_plot_count_distribution}
\title{qc plots to assess the effect of background subtraction}
\usage{
st_geomx_plot_count_distribution(ds, aes_y, aes_fill, ridge_scale = 5)
}
\arguments{
\item{aes_y}{string}

\item{aes_fill}{string}

\item{list}{dataset}
}
\value{
ggplot object
}
\description{
qc plots to assess the effect of background subtraction
}
