% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{st_geomx_plot_gene_filter}
\alias{st_geomx_plot_gene_filter}
\title{Set of plots to guide gene filtering}
\usage{
st_geomx_plot_gene_filter(ds, min_frac_expr = 0)
}
\arguments{
\item{ds}{list dataset}

\item{min_frac_expr}{number threshold range (0-1.0)}
}
\value{
ggplot object
}
\description{
Set of plots to guide gene filtering
}
