% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{st_geomx_process_input}
\alias{st_geomx_process_input}
\title{Process input data}
\usage{
st_geomx_process_input(ds, bg_lod_quantile)
}
\arguments{
\item{ds}{list produced by st_geomx_read}

\item{bg_lod_quantile}{number (0.0-1.0)}
}
\value{
list
}
\description{
Process input data
}
