% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{st_geomx_set_thresholds}
\alias{st_geomx_set_thresholds}
\title{Apply user-defined thresholds to AOIs in preparation of filtering}
\usage{
st_geomx_set_thresholds(
  ds,
  min_counts = 0,
  min_auc = 0.5,
  aoi_min_frac_expr = 0,
  gene_min_frac_expr = 0
)
}
\arguments{
\item{ds}{list produced by st_geomx_merge_probes}
}
\value{
list dataset
}
\description{
This function does not perform filtering. It sets the 'keep' column in the
sample and gene metadata allowing users to generate plots of different
thresholds.
}
\details{
Sequencing depth and signal-to-noise ratio can vary dramatically across
experiments. The default parameter values will not perform any filtering.
We recommend using the plotting functions to aid in choosing
appropriate parameter values.
}
