% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{calc_frac_expr}
\alias{calc_frac_expr}
\title{Compute frac expressed genes above threshold}
\usage{
calc_frac_expr(meta, counts, bg_lod)
}
\arguments{
\item{meta}{tibble count metadata}

\item{counts}{tibble counts}

\item{bg_lod}{vector of limit of detection thresholds per sample}
}
\value{
list
}
\description{
Compute frac expressed genes above threshold
}
