% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{normalize_quantile}
\alias{normalize_quantile}
\title{subtract background noise from each sample}
\usage{
normalize_quantile(x)
}
\arguments{
\item{x}{count matrix}

\item{offset}{vector of bg noise levels per sample}

\item{min.count}{minimum count threshold}
}
\value{
matrix of background-subtracted counts
}
\description{
subtract background noise from each sample
}
