% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgeomx.R
\name{st_geomx_normalize}
\alias{st_geomx_normalize}
\title{Normalization}
\usage{
st_geomx_normalize(ds, bgcorrect = TRUE, method = c("qn", "rle", "cpm"))
}
\arguments{
\item{method}{string normalization method}

\item{x}{matrix of counts}

\item{bgsub}{bool perform background subtraction}
}
\value{
matrix of normalized counts
}
\description{
background subtraction and scale/quantile normalization
}
